feat: manual-max-attempts 수동 재시도 한도 적용
- retryDead()에 수동 재시도 한도 검증 추가 (자동 재시도 소진 후 허용 횟수 제한)
- countByNotificationId 반환 타입 int → long 수정 (COUNT() Long 반환)
- findByIdWithInApp() 추가: open-in-view=false 환경에서 LAZY 연관 안전 로딩

+++ src/main/java/com/example/notification/infrastructure/repository/NotificationAttemptRepository.java
@@ -10,5 +10,5 @@ import java.util.UUID;
 public interface NotificationAttemptRepository extends JpaRepository<NotificationAttempt, UUID> {
 
     @Query("SELECT COUNT(a) FROM NotificationAttempt a WHERE a.notificationId = :notificationId")
-    int countByNotificationId(@Param("notificationId") UUID notificationId);
+    long countByNotificationId(@Param("notificationId") UUID notificationId);
 }
